clean up confinement for  production
