+++ packages/gnat/project.rs.xml
@@ -1,0 +1,46 @@
+<?xml version='1.0' encoding='utf-8'?>
+<Project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="https://raw.githubusercontent.com/Riverscapes/Program/master/Project/XSD/V1/Project.xsd">
+  <Name>Confinement For HUC 17060304</Name>
+  <ProjectType>Confinement</ProjectType>
+  <MetaData>
+        <Meta name="ModelVersion">0.0.6</Meta>
+        <Meta name="dateCreated">2020-10-23T01:42:01.788665</Meta>
+        <Meta name="HUC8">17060304</Meta>
+        <Meta name="Watershed">Middle Fork Clearwater</Meta>
+  </MetaData>
+  <Realizations>
+    <Confinement id="CONFINEMENT1" dateCreated="" guid="" productVersion="">
+      <Name>Valley Bottom Confinement for HUC 17060304</Name>
+      <MetaData>
+        <Meta name="ConfinementType">ValleyBottom</Meta> <!-- e.g. ValleyBottom, Anthropogenic, Composite-->
+      </MetaData>
+      <Inputs>
+        <Vector guid="" id="FLOWLINES">
+            <Name>Flowlines</Name>
+            <Path>inputs\flowlines.shp</Path>
+            <MetaData>
+                <Meta name="BufferField">BFWidth</Meta>
+            </MetaData>
+        </Vector>
+        <Vector guid="" id="VBET_50">
+            <Name>Confining Polygon</Name>
+            <Path>inputs\vbet_50.shp</Path>
+        </Vector>
+      </Inputs>
+      <Outputs>
+        <GeopackageVector guid="" id="ConfinementRaw">
+          <Name>Confinement Raw</Name>
+          <Path>Confinement.gpkg\main.Confinement_Raw</Path>
+        </GeopackageVector>
+        <GeopackageVector guid="" id="ConfiningMargins">
+          <Name>Confining Margins</Name>
+          <Path>Confinement.gpkg\main.Confining_Margins</Path>
+        </GeopackageVector>
+        <GeopackageVector guid="" id="ConfinementRato">
+            <Name>Confinement Ratio</Name>
+            <Path>Confinement.gpkg\main.Confinement_Ratio</Path>
+        </GeopackageVector>
+      </Outputs>
+    </Confinement>
+  </Realizations>
+</Project>
